+++ react/components/EXPERIMENTAL_Table/FilterBar.tsx
@@ -1,17 +1,15 @@
 import React, { FC } from 'react'
 import { InferProps } from 'prop-types'
-import classNames from 'classnames'
 
 import FilterBarBase, { filterBarPropTypes } from '../FilterBar/index.js'
 import { NAMESPACES, ORDER_CLASSNAMES } from './constants'
+import ActionBar from './ActionBar'
 
 const FilterBar: FC<FilterBarProps> = props => {
   return (
-    <div
-      id={NAMESPACES.FILTER_BAR}
-      className={classNames('mb5', ORDER_CLASSNAMES.FILTER_BAR)}>
+    <ActionBar id={NAMESPACES.FILTER_BAR} order={ORDER_CLASSNAMES.FILTER_BAR}>
       <FilterBarBase {...props} />
-    </div>
+    </ActionBar>
   )
 }