+++ react/components/EXPERIMENTAL_Table/constants.ts
@@ -15,7 +15,8 @@ export const ORDER_CLASSNAMES = {
   },
   FILTER_BAR: 'order-1',
   TOTALIZER: 'order-2',
-  BULK: 'order-3',
+  GENERIC_ACTION_BAR: 'order-3',
+  BULK: 'order-4',
   BULK_CHILD: {
     ACTIONS: 'order-0',
     ACTIONS_CHILD: {
@@ -25,8 +26,8 @@ export const ORDER_CLASSNAMES = {
     POSITION_FIXER: 'order-1',
     RIGHT: 'order-2',
   },
-  TABLE: 'order-4',
-  PAGINATION: 'order-5',
+  TABLE: 'order-5',
+  PAGINATION: 'order-6',
 }
 
 export const NAMESPACE_PREFIX = 'vtex-table-v2'
@@ -52,4 +53,5 @@ export const NAMESPACES = {
   BULK_ACTIONS: `${NAMESPACE_PREFIX}__bulk-actions`,
   CHECKBOX: `${NAMESPACE_PREFIX}__checkbox`,
   PAGINATION: `${NAMESPACE_PREFIX}__pagination`,
+  GENERIC_ACTION_BAR: `${NAMESPACE_PREFIX}__action-bar`,
 }